ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
pymongo==4.6.1
motor==3.3.2
redis==5.0.1
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.1
pydantic-settings==2.1.0
scipy==1.11.4
